refactor: rm Avatars directory
