+++ packages/bezier-react/src/components/Avatar/Avatar.module.scss
