+++ packages/bezier-react/src/components/Avatar/Avatar.stories.tsx
