+++ packages/bezier-react/src/components/Avatar/Avatar.test.tsx
