+++ packages/bezier-react/src/components/Avatar/Avatar.types.ts
