+++ packages/bezier-react/src/components/Avatar/useProgressiveImage.test.ts
