+++ packages/bezier-react/src/components/Avatar/useProgressiveImage.ts
