+++ packages/bezier-react/src/components/AvatarGroup/AvatarGroup.module.scss
