+++ packages/bezier-react/src/components/AvatarGroup/__mocks__/avatarList.ts
