+++ packages/bezier-react/src/components/AvatarGroup/index.ts
