--- packages/bezier-react/src/components/Avatars/Avatars.mdx
@@ -1,243 +1,0 @@
-import {
-  ArgsTable,
-  Canvas,
-  Story,
-} from '@storybook/addon-docs'
-import {
-  Avatar,
-} from '~/src/components/Avatars/Avatar'
-import {
-  AvatarGroup,
-} from '~/src/components/Avatars/AvatarGroup'
-import {
-  CheckableAvatar,
-} from '~/src/components/Avatars/CheckableAvatar'
-
-# Avatars
-
-## Overview
-
-프로필 이미지를 가지는 객체를 표현하기 위해 사용합니다.
-
-<Story id="components-avatars--overview" />
-
-### Smoothed corners
-
-눈썰미가 좋으신 분들은 알아보셨겠지만, **Avatar의 테두리는 완전한 원형이 아닙니다!**
-
-CSS의 `border-radius` property를 이용하면 이러한 테두리를 얻을 수 없습니다.
-[Worklet Web API](https://developer.mozilla.org/en-US/docs/Web/API/Worklet)를 사용하여, Avatar 컴포넌트만의 테두리 곡선을 직접 그려냅니다!
-Worklet API를 지원하는 브라우저에서 단순한 원형 테두리가 아닌 bezier-react만의 Avatar 형태를 볼 수 있게 됩니다.
-
-구현 방법에 대해서는 `src/features/SmoothCornersFeature` 안쪽 소스 코드를 확인해보세요 :)
-
-## Usage
-
-### Avatar
-
-Avatar를 사용하기 위한 최소한의 정보는 `name`과 `avatarUrl`입니다.
-`name`은 Avatar에 대한 label이 됩니다. `avatarUrl`은 Avatar 컴포넌트로 표현할 이미지의 리소스 URL입니다.
-
-<Story id="components-avatars--usage-basic" />
-
-```tsx
-<Avatar name="<name>" avatarUrl="<avatar-url>" />
-
-// Example
-<Avatar
-  name="Channel.io"
-  avatarUrl="https://cf.channel.io/thumb/200x200/pub-file/1/606d87d059a6093594c0/ch-symbol-filled-smiley-bg.png"
-/>
-```
-
-`avatarUrl` prop으로 주어진 resource URL의 이미지를 불러오는 동안에는, fallback 이미지를 보여줍니다.
-리소스 이미지를 불러오는 데 실패했을 때도 fallback 이미지를 보여줍니다.
-Fallback 이미지는 bezier-react에서 기본적으로 제공하나, `fallbackUrl` prop을 통해 직접 지정하는 것도 가능합니다.
-
-Avatar 컴포넌트에 대해 더 알아보려면 [스토리](/story/components-avatars-avatar--primary)를 살펴보세요.
-
-### Avatar with preset status
-
-Avatar 우측 하단에 컴포넌트를 추가하여, 아바타로 표현하고자 하는 객체의 상태를 보여줍니다.
-온라인, 오프라인, 방해금지 등의 상태를 보여줄 수 있습니다.
-
-Preset status를 사용할 때는 `status` prop에 `StatusType` enum의 값을 지정합니다.
-
-<Story id="components-avatars--usage-preset-status" />
-
-```tsx
-<Avatar
-  name="Channel.io"
-  avatarUrl="https://cf.channel.io/thumb/200x200/pub-file/1/606d87d059a6093594c0/ch-symbol-filled-smiley-bg.png"
-  status={StatusType.Online}
-/>
-```
-
-Avatar 크기가 달라져도 preset status 컴포넌트의 크기는 변하지 않습니다.
-커스텀 status 컴포넌트를 사용할 때도 유의해야 합니다.
-
-<Story id="components-avatars--usage-preset-status-with-size" />
-
-Preset status에 대해 더 알아보려면 [스토리](/story/components-status--primary)를 살펴보세요.
-
-### Avatar with custom status
-
-Preset status 이외에도, 원하는 컴포넌트를 status로 지정할 수 있습니다.
-외부 서비스와의 연동 상태를 표현하거나, 이모지를 사용하여 상태를 표현할 수 있습니다.
-
-커스텀 컴포넌트를 status로 사용할 때는, `<Avatar>` 컴포넌트의 **children**으로 지정합니다.
-
-<Story id="components-avatars--usage-custom-status" />
-
-```tsx
-<Avatar
-  name="Channel.io"
-  avatarUrl="https://cf.channel.io/thumb/200x200/pub-file/1/606d87d059a6093594c0/ch-symbol-filled-smiley-bg.png"
->
-  <Emoji // <- status 자리에 들어갈 컴포넌트를 children으로 지정합니다.
-    name="smile"
-    avatarUrl="..."
-  />
-</Avatar>
-```
-
-다른 Avatar 컴포넌트를 status로 사용할 때는, `showBorder` prop을 `true`로 지정하여 원본 Avatar와의 대비를 주어야 합니다.
-
-<Story id="components-avatars--usage-custom-status-avatar" />
-
-```tsx
-<Avatar
-  name="Channel.io"
-  avatarUrl="https://cf.channel.io/thumb/200x200/pub-file/1/606d87d059a6093594c0/ch-symbol-filled-smiley-bg.png"
->
-  <Avatar
-    name="status"
-    avatarUrl="https://bit.ly/code-beast"
-    showBorder // <-- showBorder prop을 지정하면, 원본 Avatar와의 대비가 보입니다.
-  />
-</Avatar>
-```
-
-### Disabled avatar
-
-`disabled` prop을 사용하여, 비활성화 상태의 Avatar를 표현합니다.
-
-<Story id="components-avatars--usage-disabled" />
-
-```tsx
-<Avatar
-  name="Channel.io"
-  avatarUrl="https://cf.channel.io/thumb/200x200/pub-file/1/606d87d059a6093594c0/ch-symbol-filled-smiley-bg.png"
-  disabled
-/>
-```
-
-### Checkable avatar
-
-Avatar 컴포넌트를 통해 표현되는 객체들이 있을 때, 객체들의 목록에서 선택하는 UX를 표현하기 위해 checkable avatar를 사용합니다.
-
-Checkable avatar는 선택 여부를 toggle할 수 있으며, 선택한 상태에서는 이미지 대신 체크 아이콘이 표시됩니다.
-
-<Story id="components-avatars--usage-checkable-avatar" />
-
-Avatar 컴포넌트와 유사하지만, `isChecked` 와 `isCheckable` prop을 통해 체크 상태를 표현할 수 있는 `CheckableAvatar` 컴포넌트를 사용하세요.
-
-```tsx
-<CheckableAvatar
-  name="Channel.io"
-  avatarUrl="https://cf.channel.io/thumb/200x200/pub-file/1/606d87d059a6093594c0/ch-symbol-filled-smiley-bg.png"
-  isChecked
-/>
-```
-
-Checkable avatar에 대해 더 알아보려면 [스토리](/story/components-avatars-checkableavatar--primary)를 살펴보세요.
-
-### Multiple avatars (AvatarGroup)
-
-Avatar 목록을 표현하기 위해서는 AvatarGroup 컴포넌트를 사용합니다.
-
-목록에 펼쳐놓을 최대 개수를 `max` prop으로 지정하고, 이보다 긴 목록은 두 가지 ellipsis 방식을 이용하여 표현합니다.
-Ellipsis 방법은 `ellipsisType` prop을 통해 지정합니다.
-AvatarGroup의 children으로 목록으로 표현할 Avatar 컴포넌트를 전달합니다.
-
-<Story id="components-avatars--usage-group-ellipsis" />
-
-```tsx
-<AvatarGroup
-  max={4}
-  ellipsisType={AvatarGroupEllipsisType.Icon}
->
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-</AvatarGroup>
-
-
-<AvatarGroup
-  max={4}
-  ellipsisType={AvatarGroupEllipsisType.Count}
->
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-</AvatarGroup>
-```
-
-AvatarGroup으로 표현하는 Avatar 목록은 같은 크기나 border property를 가져야 하기 때문에,
-Avatar 각각에 property를 지정하기보다는 AvatarGroup 컴포넌트에 한 번에 property를 지정하는 것이 best practice입니다.
-
-```tsx
-// DON'T ❌
-<AvatarGroup ... >
-  <Avatar size={AvatarSize.Size24} ... />
-  <Avatar size={AvatarSize.Size24} ... />
-  <Avatar size={AvatarSize.Size24} ... />
-</AvatarGroup>
-
-// DO ✅
-<AvatarGroup size={AvatarSize.Size24} ... >
-  <Avatar ... />
-  <Avatar ... />
-  <Avatar ... />
-</AvatarGroup>
-```
-
-Group 내 Avatar 간의 간격을 조절하기 위해서는 `spacing` prop을 사용하세요.
-`spacing`이 음수가 되는 경우 (즉, Avatar 간에 겹치는 경우) 또한 종종 사용되는 UX입니다.
-
-<Story id="components-avatars--usage-group-spacing" />
-
-Avatar group에 대해 더 알아보려면 [스토리](/story/components-avatars-avatargroup--primary)를 살펴보세요.
-
-## Variants
-
-### Size
-
-20x20 부터 120x120 까지, preset size가 존재합니다.
-기본 크기는 24x24 입니다.
-
-<Story id="components-avatars--variants-size" />
-
-## API
-
-### Avatar
-
-<ArgsTable of={Avatar} />
-
-### AvatarGroup
-
-<ArgsTable of={AvatarGroup} />
-
-### CheckableAvatar
-
-<ArgsTable of={CheckableAvatar} />
-
-## Version
-
-- Available since 0.3.21
