--- packages/bezier-react/src/components/Avatars/useAvatarRadiusToken.ts
@@ -1,5 +1,0 @@
-import { useToken } from '~/src/components/ThemeProvider'
-
-export function useAvatarRadiusToken() {
-  return useToken().global.radius['radius-42-p']
-}
