+++ packages/bezier-react/src/components/CheckableAvatar/CheckableAvatar.test.tsx
