+++ packages/bezier-react/src/components/CheckableAvatar/index.ts
